requirements.txt
absl-py==2.2.2
boto3==1.38.23
botocore==1.38.23
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
fonttools==4.58.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
huggingface-hub==0.24.7
idna==3.7
Jinja2==3.1.6
jmespath==1.0.1
kagglehub==0.3.12
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pillow_heif==0.22.0
protobuf==6.31.0
psutil==7.0.0
py-cpuinfo==9.0.0
pybboxes==0.1.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.64
s3transfer==0.13.0
sahi==0.11.23
scipy==1.15.3
seaborn==0.13.2
setuptools==80.8.0
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.145
ultralytics-thop==2.0.14
urllib3==2.4.0
Werkzeug==3.1.3
yolov5==7.0.14
